feat: 회원가입 API 인스턴스 생성 및 연결
