--- src/types/sign.ts
@@ -1,8 +1,0 @@
-/** 2023/08/15 - 회원가입 폼 양식 Type - sineTlsl */
-export interface SignUpFormType {
-  id: string;
-  email: string;
-  name: string;
-  password: string;
-  passwordConfirm: string;
-}
